+++ src/worker/mod.rs
@@ -1,3 +1,4 @@
+mod errors;
 mod impl_coordinator_channel;
 mod impl_execute_task;
 mod session_builder;
@@ -8,7 +9,8 @@ pub(crate) mod test_utils;
 mod worker_connection_pool;
 mod worker_service;
 
-pub(crate) use impl_execute_task::UnregisteredTaskDuringDrain;
+#[cfg(feature = "grpc")]
+pub(crate) use errors::execute_task_error_to_tonic_status;
 pub(crate) use single_write_multi_read::SingleWriteMultiRead;
 pub(crate) use worker_connection_pool::WorkerConnectionPool;
